+++ database/migrations/2022_11_23_211150_create_product_in_details_table.php
@@ -27,7 +27,7 @@ class CreateProductInDetailsTable extends Migration
                 ->restrictOnDelete();
 
             $table->float('qty');
-            $table->decimal('price');
+            $table->float('price');
             $table->unique(['product_id', 'product_in_id']);
         });
